pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "robotics-datasets-hub"
version = "0.1.0"
description = "One-command download, convert, and visualize curated AI/Robotics datasets"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [{ name = "rsasaki0109" }]
keywords = ["robotics", "datasets", "autonomous-driving", "navigation", "slam"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "pyyaml>=6.0",
    "huggingface-hub>=0.20",
    "gdown>=5.0",
    "requests>=2.31",
    "tqdm>=4.66",
    "matplotlib>=3.8",
    "numpy>=1.26",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]
viz3d = [
    "open3d>=0.18",
    "plotly>=5.18",
]

[project.scripts]
rdh = "rdh.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/rdh"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
